Add files via upload
Blackjack, formerly also Black Jack and Vingt-Un, is the American member of a global family of banking games known as Twenty-One, whose relatives include the British game of Pontoon and the European game, Vingt-et-Un. It is a comparing card game between one or more players and a dealer, where each player, in turn, competes against the dealer. Players do not compete against each other. It is played with one or more decks of 52 cards and is the most widely played casino banking game in the world.

Rules
Players are each dealt two cards, individually facing up or down depending on the casino and the table. In the U.S., the dealer is also dealt two cards, normally one up (exposed) and one down (hidden). In most other countries, the dealer only receives one card face up. The value of cards two through ten is their pip value (2 through 10). Face cards (Jack, Queen, and King) are all worth ten. Aces can be worth one or eleven. A hand's value is the sum of the card values. Players are allowed to draw additional cards to improve their hands. A hand with an ace valued as 11 is called "soft", meaning that the hand will be guaranteed to not score more than 21 by taking an additional card. The value of the ace will become one to prevent the hand from exceeding 21. Otherwise, the hand is called "hard".

Once all the players have completed their hands, it is the dealer's turn. The dealer hand will not be completed if all players have either exceeded the total of 21 or received blackjacks. The dealer then reveals the hidden card and must draw cards, one by one, until the cards total up to 17 points. At 17 points or higher, the dealer must stop. (At most tables the dealer also hits on a "soft" 17, i.e. a hand containing an ace and one or more other cards totaling six.) Players bet on the basis that they will individually have better hands than the dealer. The better hand is the hand where the sum of the card values is closer to 21 without exceeding 21. The detailed outcome of the hand follows:

If the player is dealt an Ace and a ten-value card (called a "blackjack" or "natural"), and the dealer does not, the player wins and usually receives a bonus.
If the player exceeds a sum of 21 ("busts"), the player loses, even if the dealer also exceeds 21.
If the dealer exceeds 21 ("busts") and the player does not, the player wins.
If the player attains a final sum higher than the dealer and does not bust, the player wins.
If both dealer and player receive a blackjack or any other hands with the same sum, this will be called a "push" and no one wins.
Blackjack has been estimated to have over 100 rule variations. Since the 1960s, blackjack has been a high-profile target of advantage players, particularly card counters, who track the profile of cards that have been dealt and adapt their wagers and playing strategies accordingly. In response, casinos have introduced counter-measures that can increase the difficulty of advantage play.

Blackjack has inspired other casino games, including Spanish 21 and pontoon.